requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
openai
unittest
